+++ lib/src/codecs/base64.dart
@@ -12,11 +12,11 @@ const _base64Encoding = [
 ];
 
 const _base64Decoding = [
-  -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, //
-  -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-  -1, -1, -1, -1, -1, 62, -1, -1, -1, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60,
-  61, -1, -1, -1, -1, -1, -1, -1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
-  14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1, -1, -1, -1, 26,
+  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, //
+  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2,
+  -2, -2, -2, -2, -2, 62, -2, -2, -2, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60,
+  61, -2, -2, -2, -1, -2, -2, -2, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
+  14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -2, -2, -2, -2, -2, -2, 26,
   27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45,
   46, 47, 48, 49, 50, 51
 ];
@@ -24,6 +24,12 @@ const _base64Decoding = [
 class B64Codec extends Uint8Codec {
   @override
   final Uint8Encoder encoder;
+
+  @override
+  final decoder = const Uint8Decoder(
+    bits: 6,
+    alphabet: _base64Decoding,
+  );
 
   const B64Codec()
       : encoder = const Uint8Encoder(
@@ -37,12 +43,6 @@ class B64Codec extends Uint8Codec {
           padding: 61,
           alphabet: _base64Encoding,
         );
-
-  @override
-  final decoder = const Uint8Decoder(
-    bits: 6,
-    alphabet: _base64Decoding,
-  );
 }
 
 /// Codec to encode and decode an iterable of 8-bit integers to 6-bit Base64