+++ lib/src/codecs_base.dart
@@ -2,6 +2,7 @@
 // All rights reserved. Check LICENSE file for details.
 
 export 'codecs/ascii.dart' hide ASCIICodec;
+export 'codecs/base2.dart' hide BinaryCodec;
 export 'codecs/base16.dart' hide B16Codec;
 export 'codecs/base32.dart' hide B32Codec;
 export 'codecs/base64.dart' hide B64Codec;